doc comments, clean up, and fix mentat macro
