+++ crates/mentat-server/src/api/account.rs
@@ -6,7 +6,7 @@ use super::*;
 
 /// AccountAPIServicer defines the api actions for the AccountAPI service
 #[axum::async_trait]
-pub trait AccountApi: Clone + Debug + Send + Sync {
+pub trait AccountApi: Clone + Debug + Default + Send + Sync {
     /// the caller used to interact with the underlying node
     type NodeCaller: Clone + Debug + Send + Sync + 'static;
