+++ crates/mentat-server/src/api/block.rs
@@ -6,7 +6,7 @@ use super::*;
 
 /// BlockAPIServicer defines the api actions for the BlockAPI service
 #[axum::async_trait]
-pub trait BlockApi: Clone + Debug + Send + Sync {
+pub trait BlockApi: Clone + Debug + Default + Send + Sync {
     /// the caller used to interact with the underlying node
     type NodeCaller: Clone + Debug + Send + Sync + 'static;
