+++ crates/mentat-server/src/api/call.rs
@@ -6,7 +6,7 @@ use super::*;
 
 /// CallAPIServicer defines the api actions for the CallAPI service
 #[axum::async_trait]
-pub trait CallApi: Clone + Debug + Send + Sync {
+pub trait CallApi: Clone + Debug + Default + Send + Sync {
     /// the caller used to interact with the underlying node
     type NodeCaller: Clone + Debug + Send + Sync + 'static;
