+++ crates/mentat-server/src/api/mempool.rs
@@ -6,7 +6,7 @@ use super::*;
 
 /// MempoolAPIServicer defines the api actions for the MempoolAPI service
 #[axum::async_trait]
-pub trait MempoolApi: Clone + Debug + Send + Sync {
+pub trait MempoolApi: Clone + Debug + Default + Send + Sync {
     /// the caller used to interact with the underlying node
     type NodeCaller: Clone + Debug + Send + Sync + 'static;
