+++ crates/mentat-server/src/api/network.rs
@@ -6,7 +6,7 @@ use super::*;
 
 /// NetworkAPIServicer defines the api actions for the NetworkAPI service
 #[axum::async_trait]
-pub trait NetworkApi: Clone + Debug + Send + Sync {
+pub trait NetworkApi: Clone + Debug + Default + Send + Sync {
     /// the caller used to interact with the underlying node
     type NodeCaller: Clone + Debug + Send + Sync + 'static;
