+++ crates/mentat-server/src/api/search.rs
@@ -6,7 +6,7 @@ use super::*;
 
 /// SearchAPIServicer defines the api actions for the SearchAPI service
 #[axum::async_trait]
-pub trait SearchApi: Clone + Debug + Send + Sync {
+pub trait SearchApi: Clone + Debug + Default + Send + Sync {
     /// the caller used to interact with the underlying node
     type NodeCaller: Clone + Debug + Send + Sync + 'static;
