+++ mentat-macros/src/lib.rs
@@ -5,7 +5,6 @@
 
 extern crate proc_macro;
 
-mod route_builder;
 mod unchecked;
 
 use proc_macro::TokenStream;
@@ -99,18 +98,12 @@ fn gen_main(
     _server_type: &Ident,
     _cache_type: Option<&Ident>,
 ) -> TokenStream2 {
-    // let routes = route_builder::build_routes(server_type, cache_type);
-
     quote!(
         use ::mentat_server::{conf::NodePid, macro_exports::tokio, sysinfo::Pid};
         #[tokio::main]
         async fn main() {
             use ::mentat_server::macro_exports::*;
             let server = #server_call;
-            // let routes = |mut app: Router<Arc<AppState<<#server_type as ServerType>::CustomConfig>>>| {
-            //     #routes
-            //     app
-            // };
             server.serve().await
         }
     )
